WIP: Introduce vhost module for LLNG
We add a new Ansible module to handle VHost configuration on
LemonLDAP::NG through the LLNG CLI.

Currently we use the CLI because the manager API don't allow to
manipulate VHost configuration. Adding new item to menu can
probably done with the REST API, we will support that later on follow up
commit.